+++ backend/cmd/internal/http/handler/websocket_routes.go
@@ -12,7 +12,7 @@ import (
 )
 
 type WebSocketService interface {
-	RegisterConnection(userID int, sessionID string, connID string, exp int64) apierror.ErrorResponse
+	RegisterConnection(userID int64, sessionID string, connID string, exp int64) apierror.ErrorResponse
 	RemoveConnection(connectionID string)
 	HandleMessage(msg *contract.IncomingSocketMessage, connID string)
 }
